diag_os_chronoms: fixes
Make treset parameter signed to enable usage like chronoms(-t), used in
diag_os_calibrate to reset previous chronoms offset.
Hack for UNIX: the "seconds * 1000" multiplications (the number of
milliseconds since 1970) overflow in 32 bits, so cast to unsigned so as to
at least ensure wrapping behaviour.

+++ scantool/diag_os_win.c
@@ -400,12 +400,12 @@ unsigned long diag_os_getms(void) {
 
 
 //millisecond stopwatch, arbitrarily resettable
-unsigned long diag_os_chronoms(unsigned long treset) {
-	static unsigned long offset=0;
+long diag_os_chronoms(long treset) {
+	static long offset=0;
 
 	offset += treset;
 
-	return (unsigned long) (GetTickCount() - offset);
+	return GetTickCount() - offset;
 }
 
 //get high resolution timestamp
